Use only push-back instead of emplace-back
emplace_back() was always called with a reference anyway, and
push_back() is more generic when using custom or 3rd-party non-STL
collections.

Partitioning can only be done in collections where the value type is the
same as the sequence's value type. If a different value type is needed,
the sequence can be mapped before partitioned.